Add ngeoGeolocationOptions
