+++ contribs/gmf/apps/desktop/index.html.ejs
@@ -252,7 +252,6 @@
         ></ngeo-displaywindow>
         <button ngeo-geolocation=""
             ngeo-geolocation-map="::mainCtrl.map"
-            ngeo-geolocation-options="::mainCtrl.geolocationOptions"
             ng-if="::mainCtrl.options.geolocalisation">
             <span class="far fa-dot-circle"></span>
         </button>
