+++ contribs/gmf/apps/desktop_alt/index.html.ejs
@@ -294,7 +294,6 @@
         ></ngeo-displaywindow>
         <button ngeo-geolocation=""
             ngeo-geolocation-map="::mainCtrl.map"
-            ngeo-geolocation-options="::mainCtrl.geolocationOptions"
             ng-if="::mainCtrl.options.geolocalisation">
             <span class="far fa-dot-circle"></span>
         </button>
