+++ contribs/gmf/apps/mobile/index.html.ejs
@@ -79,7 +79,6 @@
       </div>
       <button ngeo-geolocation=""
         ngeo-geolocation-map="::mainCtrl.map"
-        ngeo-geolocation-options="::mainCtrl.geolocationOptions"
         ng-if="::mainCtrl.options.geolocalisation">
         <span class="far fa-dot-circle"></span>
       </button>
