+++ contribs/gmf/apps/mobile_alt/index.html.ejs
@@ -73,7 +73,6 @@
       </div>
       <button ngeo-geolocation=""
         ngeo-geolocation-map="::mainCtrl.map"
-        ngeo-geolocation-options="::mainCtrl.geolocationOptions"
         ng-if="::mainCtrl.options.geolocalisation">
         <span class="far fa-dot-circle"></span>
       </button>
