+++ contribs/gmf/src/query/windowComponent.js
@@ -31,7 +31,7 @@ import ngeoQueryMapQuerent from 'ngeo/query/MapQuerent.js';
 
 import olCollection from 'ol/Collection.js';
 import {isEmpty} from 'ol/obj.js';
-import {buildStyle} from 'gmf/options.js';
+import {buildStyle} from 'ngeo/options.js';
 
 import 'jquery-ui/ui/widgets/resizable.js';
 import 'ngeo/sass/jquery-ui.scss';
